change complete-info.js to complete-info.ts
