+++ src/test/js/jan/yaku-util-test.js
@@ -15,7 +15,7 @@ limitations under the License.
  */
 import assert from 'assert';
 
-import CompleteInfo from '../../../main/js/jan/complete-info';
+import {CompleteInfo} from '../../../main/js/jan/complete-info';
 import CompletePattern from '../../../main/js/jan/complete-pattern';
 import {COMPLETE_TYPE} from '../../../main/js/jan/complete-type';
 import Hand from '../../../main/js/jan/hand';
